+++ src/main/java/com/habitpay/habitpay/global/error/CustomJwtErrorInfo.java
@@ -9,5 +9,5 @@ public enum CustomJwtErrorInfo {
     UNAUTHORIZED("invalid_token"), // 401
     FORBIDDEN("insufficient_scope"); // 403
 
-    private final String errorMessage;
+    private final String Message;
 }
