+++ src/main/java/com/habitpay/habitpay/global/handler/GlobalExceptionHandler.java
@@ -1,6 +1,6 @@
 package com.habitpay.habitpay.global.handler;
 
-import com.habitpay.habitpay.global.exception.JWT.ErrorTokenResponse;
+import com.habitpay.habitpay.global.exception.ErrorMessageResponse;
 import com.habitpay.habitpay.global.exception.JWT.CustomJwtException;
 import com.habitpay.habitpay.global.exception.PostPhoto.CustomPhotoException;
 import lombok.extern.slf4j.Slf4j;
@@ -13,18 +13,21 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
 @Slf4j
 public class GlobalExceptionHandler {
     @ExceptionHandler(CustomJwtException.class)
-    protected ResponseEntity<ErrorTokenResponse> customJwtExceptionError(CustomJwtException exception) {
+    protected ResponseEntity<ErrorMessageResponse> customJwtExceptionError(CustomJwtException exception) {
         return ResponseEntity.status(exception.getStatusCode())
-                .body(new ErrorTokenResponse(
-                        exception.getCustomJwtErrorInfo().getErrorMessage(),
-                        exception.getErrorMessage()
+                .body(new ErrorMessageResponse(
+                        exception.getCustomJwtErrorInfo().getMessage(),
+                        exception.getMessage()
                 ));
     }
 
     @ExceptionHandler(CustomPhotoException.class)
-    protected ResponseEntity<String> customPhotoExceptionError(CustomPhotoException exception) {
+    protected ResponseEntity<ErrorMessageResponse> customPhotoExceptionError(CustomPhotoException exception) {
         return ResponseEntity.status(exception.getStatusCode())
-                .body(exception.getErrorResponse() + exception.getMessage());
+                .body(new ErrorMessageResponse(
+                        exception.getErrorResponse().getMessage(),
+                        exception.getMessage()
+                ));
     }
 
     // todo : getMessage() 숨기기?
